Prepare Do_Core for better no-frame eval optimizations
This brings forward some changes from the specific binding branch that
were related to making it possible for ET_WORD to be dispatched for non
function lookups without needing a Reb_Frame.  The actual changes are
tied up in the ability to use inline functions, but this gets the
naming and behavior in sync on the pre-specific-binding branch.