Ground zero beginnings of specific binding
This begins specific binding by starting the threading of "specifiers"
through, mostly just by beginning to ensure that functions are taking
the right number of arguments.  The specifier is mostly unheeded at
this point, as generalized code has not been retrofitted to thread it
through correctly.

To make note of cases that are knowingly not threading the specifier,
the GUESSED constant is provided.  This stopgap should make it easier
to move the code forward a little bit at a time.