Do not maintain THROWN_FLAG or END_FLAG in Do_Core
The THROWN_FLAG and END_FLAG were artifacts of the DO/NEXT interface
in R3-Alpha.  It wished to return the next index, or an END or whether
the operation threw.  Since there was no structure to hold the state
of the evaluator, the way state was held was in an integer with
reserved values for the flags.

Ren-C's needs for a debugger brought about the requirement for frame
state to be reflected...but also there are optimizations possible when
a state structure is held.  (Reuse of lookahead fetches, for instance.)
This changes the internals to not do bookkeeping on THROWN_FLAG and
END_FLAG, and only expose that to those clients that don't use the
Reb_Frame API.