Speed fetch by putting va_list and eval_fetched under branch
The FETCH_NEXT_ONLY_MAYBE_END inlined function was small, but the
fundamental means of advancing the evaluator through an array.  It had
3 branches...for ordinary array fetching, for VA_LIST fetching, and
for getting a virtual value out of a "saved" slot, e.g. for the
purposes of an EVAL.

About 2-3% of evaluator time was spent just in this, and it was not
optimized for the drastically most common case of advancing the array.
This switches it so that the first test it's able to do is for whether
the common case applies.

This also solidifies the END_FLAG/THROWN_FLAG are "higher level API"
only, by putting VA_LIST_FLAG into that mix.  This means that the
indexing field in the frame is simply an index, and not an "indexor".

VALIST_FLAG is renamed to VA_LIST_FLAG, to help distinguish it from
the "DO_FLAG_HAS_VARLIST".