END signifies unspecialized slots, tidying
This changes the marker for an unspecialized slot from VOID to an end
cell.  As a result, this makes it mechanically possible to distinguish
argument slots that have been purposefully set to be "opted out of"
from a slot which is unspecialized and requires fulfillment.

To make things clearer in the code, a Specialized_Arg() function
abstracts the usages--which would make it easier if this had to change
in the future.

On the downside of the change, it limits callers using GC-visible frames
from anything but full specialization.  This is because END markers
cannot be in an array that the GC is aware of (the argument cells are
protected specially by the frames).  Generally speaking this is not
looking like a problem, since not copying frames afresh when DO-ing
them looks to be a bad idea.

Also bites the bullet and makes a boolean to track the refinement
pickups, as there was a bug in abusing the varargs subfeed.