Move function spec to the paramlist MISC field
It became necessary to modify metadata about a function after its
creation.  For instance, the descriptions of functions that are
anonymous and made by generators--to avoid each such generator from
having to come up with its own "spec" facility.

Because of the way functions worked, it was not possible to change these
specs without changing the values themselves--as the `spec` field was
in the canon value.  This uses the trick of moving the spec into the
`misc` field of the REBSER node of the array of the paramlist.

Mechanically this means that the VAL_FUNC_SPEC macro cannot be used
until after the paramlist has been established in the function value.
So this required some rote reordering.

Besides being useful for the original purpose, freeing up an extra
pointer in the function itself offers an optimization opportunity of
using it to put a custom dispatcher into the value cell.