Let definitional returns use freed up REBVAL slot
The freeing of the spec from the FUNCTION! REBVAL left room to put the
`exit_from` specifier, that had previously been "hacked" over top of
the root RETURN and LEAVE natives code pointer--since there was nowhere
else to put it.

In addition to there being less exceptionalism needed to handle the
corrupted function pointer, it means there's a function to run.  This
lets a small amount of special behavior move out of the giant Do_Core
routine and back into the RETURN and LEAVE natives.  This possibly also
leads toward a path where other natives (or user functions) make use
of the additional pointer in a similar way.