Split stackvars/varlist, VAL_CONTEXT_EXIT_FROM
This tweaks the previous solution to create a new keylist every time a
context needed to be made for a definitional LEAVE or RETURN, to make
it so that context values themselves can store the extra pointer.
This is to help potential generalizations of the technique if it
becomes more widespread as a user-available tool for any USE that
wants to make a definitional function available (BREAK, CONTINUE, KEEP)

To make space, the `stackvars` field was eliminated from the context...
instead accessing it as a persistent field of the frame (which already
had a pointer in the context).  While this makes access of the variables
through FRAME! objects for the user one step more indirect, it makes
the more common access and inspection in Do_Core more direct...and
eliminates the need for a separate DO_FLAG_HAS_VARLIST.