TYPE? to `<r3-legacy>`, kill STACK, move comments
The TYPE? construct as asking a question seems very close to DATATYPE?.
With deprecations to UNSET? and VALUE? already in the works, this goes
ahead and replaces it when not running in legacy mode with a message.

The STACK legacy construct was broken, and there is no goal to maintain
R3-Alpha "debugger compatibility".  So this just deletes the experiment.

This also moves some of the essays (or just uncovered essay points) into
the porting Trello, leaving behind links to help reduce clutter.
More of this work to do remains.