CONTEXT_SPEC => CONTEXT_META, add meta-of and set-meta
The R3-Alpha module system had a REB_MODULE type, which saved an extra
"spec" field inside the REBVAL.  Although it was called a spec, it
was the context behind an OBJECT! describing the various fields in the
module header.  Hence it was basically an object that had another
object linked as "meta-information", so that the fields in the header
would not interfere with the fields in the context that constituted
the module definition.

Having the feature of "associated meta-information" exclusively for
modules seemed arbitrary.  Also, the interesting idea behind the module
system is to implement its rules/protections/scoping-logic in
user-mode.  So if the only thing differentiating a MODULE! from an
object was this feature, it seemed it should be more generic.

This takes the step of making the feature more generic, as well as
making it possible to change the "meta" object associated with another
object so it affects all the instances of that object.  This is done
by moving the meta information to the misc field of the keylist
(while the keylist is already in the misc field of the varlist).

The goal of the feature is to make it more generic, and unify it with
a similar feature for FUNCTION! to store an object instead of just a
block in its description information.

With the new feature in place, it was possible to get rid of the
TO MODULE! conversion step needed when objects are being formed, which
further pushes the module system closer to being all user-mode code.