kill What-Reflector; XXX-OF and ANY-XXX? table-driven
The boot code had a problem with a table driven approach to making
lists of functions and setting them.  The problem was that the things
being set had to be top-level source SET-WORD!s in the file.  Hence
some amount of repetition was necessary to name the thing-being-set
and the parameterization for setting it.

This switches %base-defs.r to use a trick with variadics to consume
the top level set words from a "table" that is the variadic data
following the function invocation, up to an expression barrier.

This commit also gets rid of the `What_Reflector()` function and the
OF_XXX level of indirection from symbols--which were essentially just
in the way.  (Also, canon word values were not checked, so REFLECT
of a 'spec and of a 'SPEC gave different answers...)