Boot reordering, IS function, experimental PROBE
This reorganizes the boot slightly to put the %base-defs.r before the
%base-func.r.

It also adds the IS function for type checking.  IS was one of the
longstanding two-letter words that was not being used, and was
reconceived as a "non-logical" type test, that passes through or blanks:

    >> is string! "abc"
    "abc"

    >> is string! 123
    _

    >> is [string! integer!] 123
    123

Since IS is void-tolerant in its argument, this creates a nice option
for dealing with the void-returning SELECT, if one is more interested
in getting a blank.  It's an opportunity to do some type enforcement:

    if x: is string! select [a "foo" b "bar"] 'c [...]

Even if there's no assignment, the type can help serve as documentation
of the expected type(s).

Included is also an experimental tweak to PROBE, to make it sense when
it is passed a literal block using SEMIQUOTED?...and use a dialect
similar to ?? in R3-Alpha.  This dialect can also evaluate GROUP!s, and
will trap errors in those evaluations but keep running.

Adds NEITHER? and BOTH? and moves some codec stuff.