Handle throws in the middle of print EVAL, assert fix
If a PRINT did a call into a GROUP! that threw, then it didn't drop
the mold state in progress.  There was also an over-aggressive assert
which wasn't taking into account the possibility of throws or evals
when measuring the not-ended state of the end of a Do_Core().  This
keeps the assert fairly tight, just to make sure everything's
understood.