Support SEMIQUOTED? testing on variadic parameters
This is a somewhat speculative feature, which was thought to be needed
for something it turned out not to be needed for.  But it still might
be useful.  Basically, it updates the information tracked on a variadic
argument to indicate whether the last fetch of it was evaluated or not.

It's slightly tricky, because the argument slot which holds the bit
can be overwritten with another non-vararg value, which would mean
the bit is mechanically being set on a value that it doesn't apply to.
Since users aren't supposed to be aware of *how* semiquoted tests are
done, this isn't that big an issue, but it's a little puzzling to be
reading the bit off something that "isn't a vararg any more" when you're
TAKE-ing it out of some other variable.