Patch HELP and ENDable handling
The usage of END for in-progress arguments meant that as the argument
loop went by, it had to put VOID in the slots to make them valid.  The
cases for END on BAR! or end of series were not doing this, leading to
corrupt series.

This also includes patches to make HELP without arguments display the
banner under the read-only-function-body rule.