Eliminate VAL_FUNC_CLASS in favor of dispatcher
This uniformly switches all functions to have a "body" array, and
switches out the notion of how dispatch is done for functions to
being held in a value reachable in the "misc" field of the body array.

Beyond performance optimization of not needing a switch statement to
dispatch functions, this means it can be possible to revector any
function with a "hook".  That feature is not implemented here, but is
a work in progress.