Fix assertion in COPY of a MAP!
The VAL_ARRAY_XXX functions now refuse to work if the value is a MAP!
and not an ANY-ARRAY!.  In order to get the array out, one must go
through a step of VAL_MAP to get the REBMAP*, and then MAP_XXX to
extract the pairlist or the hashes.

This does that transformation in order to fix the assert.  It also
backs off the idea that maps are created with an anticipation of using
BAR! to separate items.  This feature should be removed from the
base MAKE MAP! and instead into an optimized native which is designed
to be something that a user could write.