REDESCRIBE routine for relabeling functions/args
Not necessarily its final name, but REDESCRIBE can be used to change
or remove description strings from functions.

This involves an overhaul of the code for turning specs into function
parameter lists, and relies upon the ability to add tagged "meta"
information onto functions.  The meta information takes advantage of
the FRAME! data type's ability to carry a cell of information per
function parameter while reusing the parameter list of the function
as the "keylist".