Get rid of SET_VOID_UNLESS_LEGACY_NONE
This removes the #ifdef for the behavior that routines like IF, SWITCH,
WHILE etc. would return BLANK! values instead of no value (void) if the
branches did not run.

The reason it is being removed is because new abilities for adapting the
functions from the outside are making it possible to do the adaptation
in client code.  Those changes will be in a subsequent commit.  For
this commit, however, the r3-legacy compatibility on that feature will
not be working.

(This is in a separate commit to isolate the pieces of that change.)