Replace action spec hack with generic spec hack
There was a shift to where functions do not preserve the spec they are
given, but instead "meta objects" that more structurally describe the
parameters and titling.  Yet some code in the ports depended on being
able to reach into the spec blocks for the built-in "actions" after
the fact.

This code does not seem to be currently active.  However, if the need
did come up it would have to generate a spec block back from the meta
object information.  This includes some code that does so, which will
be tested on an as-needed basis if this ever runs.  :-/