First draft of ADAPT and CHAIN, function reorganization
This commit contains some general shakeups around how function dispatch
works, and an attempt at better generalization of the process of
"inheritance" as used by HELP and REDESCRIBE.  It also tries to better
coalesce the ways in which function dispatch gets "X-ray vision" on
a function it is going to run, to properly prepare it to prepare the
frame appropriately for the underlying process--regardless of how many
specializations and adaptations have been done.

Included is an update of r3-legacy regarding the voids-to-blanks
conversion, which uses CHAIN to process void results from functions
like IF, CASE, WHILE, or PICK and produce blanks out of them.