Add ANY? and ALL? via CHAIN, XXX? descriptions
With a working REDESCRIBE, it is now possible to give descriptions to
specialized functions such as IF? and CASE?.  And with CHAIN it is
possible to make ANY? and ALL?

Removes FOUND? for a chained FIND? replacement, adding SELECT? also.
Includes a few substitutions from the mezzanine for the new ?-routines.