Eliminate straggling FOUND? calls
An errant search-and-replace missed some calls to FOUND?, which was
moved to R3-Legacy as specialized and chained alternatives like
FIND? and ANY? are better than trying to alias "NOT BLANK?" as being
generically synonymous with "Found".

(In Ren-C it isn't, so a better match would be "set?", but that's not
universal since things like FIND and ANY do return blank.  So it's
best to leave it up to their XXX? alternate forms to do the
specialization as appropriate.)