Fix IS function and add /RELAX refinement
A breaking change to the IS function made it return the value regardless
of whether the type matched.

Besides repairing that problem, this addresses a potential source of
user errors, by making it so that IS will not type test for LOGIC! or
BLANK! by default.  This would prevent users from getting as confused
if they write something like:

    value: false
    if is [integer! logic!] value [
        print "expected this to run, why doesn't it?"
    ]

That will make the IF IS pattern safer.

Also, small change to refinement arguments to mark them as all accepting
voids.