Streamline SWITCH and CASE, make ANY and ALL safer
This commit pares down the comments and legacy behaviors of SWITCH and
CASE, with the idea of moving the relevant legacy switches out into
the r3-legacy module using the new adaptation and specialization
facilities.

It also reduces the amount of copying of cells back and forth that CASE
has to do.

The nebulously-named PUSH_CALL_UNLESS_END is re-introduced as a
"SAFE_ENUMERATOR", and used in all four functions to avoid crashes if
there are changes to the block being enumerated while running.