Updates to ADAPT and "underlying" function mechanic
The "function composers" are a work-in-progress, and the exact shape
of how the composition is to be handled efficiently (or at all) has
been "emerging".

This takes some further steps in understanding the order and nature of
operations in a function chain--and how that interacts also with APPLY
and DO of a FRAME!.  It cements the idea that the frame is always
pushed w.r.t. the single "underlying" function, to which any adaptations
or applies must bind, even if it has more parameters than the spec of
the function they were designed for.

Included is an adaptation of SWITCH that delivers error messages on
usages that would have changed from R3-Alpha.  This is in place of the
excessive commenting and #ifdefs that were taken out in the previous
commit.