Initial experimental HIJACK implementation
This adds the ability to replace a function's implementation with
another's, as well as to receive a value that will "look and feel"
like the original function value.

It relies on some trickery to do what it does--which will trip up
various aspects of the system.  One thing that definitely will not work
is that an attempt to call a function through DO of a frame that was
created with MAKE FRAME! on the "look and feel" value given back from
HIJACK will route through the hijacker--not the original code.

There may be issues related to changing a function's dispatcher out
from under it, which will require thinking to nail down.  But this
commits the feature so that it can be merged into the specific
binding branch to see what the effects are there.