Synchronize string token table with lexer errors
The %strings.r file contains a string corresponding to each lexer token
state.  In some cases, this table is consulted to turn a token state
into an error.

The token state enumeration had drifted out of sync with the string
constants, leading to confusing messages picked from random locations
in the string table.  This addresses that immediate need--though
building this table from an abstract specification so it couldn't
get out of sync would be a better idea.