Patch recursive MAP! molding bug
On a case like:

    a: map[] a/1: a probe a

The guarding code that checked to make sure an in-progress array wasn't
molded again was trying to use VAL_ARRAY to extract the array from a
map instead of MAP_PAIRLIST.  Asserts caught this to make sure the
caller knew what they were doing.

This uses MAP_PAIRLIST for the extraction in the map case.