Further unify routines for making paramlists from specs
One idea in Ren-C's formulation of functions was to change it so that
MAKE FUNCTION! did not recognize any specific English keywords, but
was cued entirely by the types of words, positions of blocks, etc.
This meant that transformations were done, e.g. [<opt> integer!] would
be turned into [_ integer!]--to use a BLANK! literal instead of a
word for <opt>.  This meant FUNC and PROC were layers above
MAKE FUNCTION! that any user could create--picking different words for
RETURN and LEAVE, or making any decision differently.

Yet to make FUNC and PROC efficient, they were hand-coded as natives.
Their processing ultimately had the goal of making a paramlist--just
as MAKE FUNCTION! did.  However, they recognized more patterns.

The original implementation for FUNC and PROC actually copied the
spec given by the user and transformed it to something that MAKE
FUNCTION! could read.  In addition to being an intermediate step with
additional cost to process and GC, this created a lossy situation.  The
MAKE FUNCTION! code then would save the spec in its mutated form...so
HELP would not see what the user wrote, but what FUNC or PROC wrote
for MAKE FUNCTION!.

With the change to functions not storing their specs, and generators
having more free reign to write meta information onto the function
value--the balanced changed.  It no longer makes sense to generate an
intermediate deep copy of the spec.  Instead, this weaves together a
single Make_Paramlist loop that does all the work--conditional on a
flag of whether it should be aware of the keywords or not.