Routine fixes for %gtk.r and %qsort.r
With the various shake-ups in the function code, the gtk.r and qsort.r
tests were not running.

This patches the code to make them run again.  It also changes the test
for if a function is either a routine or a callback to use the name "RIN"
instead of the name "Routine" to help avoid confusion.