Do not access uninitialized error slots w/CTX_VAR
The CTX_VAR and CTX_KEY accessors check to make sure that the value is
in a valid state.  So when the error slots are reading from invalid
spots past the end of initialized content, they should do it by
incrementing the returned pointer for the current last slot instead
of trying to pass in the index of the past-last slot.