Integrate core-tests fork directly into project
The rebolsource/rebol-test repository was initially made as a submodule
of Ren-C.  Over the long run this turned out to just make matters
more complicated.  Given the divergence from the content of the original
repository--as well as the lack of updates to the original--it is
not worth it to keep them separate.

Besides moving the files into the repository, this includes a quick
conversion of the documentation into MarkDown as README.md, with some
reorganization and making it a bit shorter.