Support saving generated files to out-of-source folder
A function parse-args is added to common.r. The arguments are supposed to
be in the form of "name=value". Currently, only names that are used are
OS_ID and OUTDIR. OS_ID is used for bootstrapping for a new platform. OUTDIR
specifies the root directory where the generated files will be saved.

e.g.
${REBOL} ${TOOLS_DIR}/make-boot.r OS_ID=${R3_OS_ID} OUTDIR=${TOP_GENERATED_SRC_DIR}