Initialize eval_type for safe enumerators
A "safe enumerator" is currently just implemented as pushing a frame.
However, the `eval_type` frame wasn't initialized, and could be read
by the garbage collector, or by things looking up the stack to test
to see if a punctuator was trying to be used as an argument.

Finding a way to reduce the number of fields initialized in each
safe enumerator push, e.g. with a "reduced field validity" DO_FLAG,
is likely a better solution--but for now this takes care of the problem.