Revisit "NewPrint", /EVAL and newline behavior
This updates the NewPrint experiment to reign in the code a little bit,
and to implement the safer behavior of not printing blocks that are
products of evaluation without an /EVAL switch.

It also gets rid of the idea of BAR! being used to express newlines,
so that | can be used as a more general term separator.  This doesn't
prevent it notationally from being placed in locations where newlines
are, it just doesn't generate them itself.

This also introduces the experiment of placing newlines between
sequential string literals, if they were not products of evaluation.