Test suite updates for MAKE/TO changes
The merging of the MAKE and TO changes carried along the idea that all
make definitions for arrays had to be length 2, with the series and
the index.  The branch of the tests with these changes did not get
incorporated when the test suite was added to the %tests directory.

Also updates some molding and loading code to bridge compatibility for
some tests that the MAKE/TO branch had just thrown out.