Unify frame's `lookback` flag with the `eval_type`
There was a separate flag for tracking if a function was "lookback" that
needed to be managed by the frame.  This flag could only be valid if
an evaluation of a function was being performed, and when the function
was not being evaluated any longer it would have to be cleared.

It is more efficient (and easier to track) if this is merely another
evaluation state.  e.g. ET_LOOKBACK implies ET_FUNCTION and the lookback
flag.  That way, when the eval_type is updated, it takes care of
overwriting the flag.

Additionally, by making ET_FUNCTION 0 and ET_LOOKBACK 1, it is possible
to use the reading of the lookback flag by Get_Var_Core to do double
duty by setting up the evaluation type...while still returning something
other clients can use as a boolean.

This reorganizes the comments in the frame also, to update them a bit.
The change was motivated also by wanting to get rid of the non-standard
REBOOL in the struct layout for frames, which Ren-Cpp wants to be able
to read.