Fix sequence point problem in `read/lines`
Macros for initialization like Val_Init_Block are now "trashing" the
output slot.  This concept was originally introduced to deal with the
type correctness issue for specific binding (e.g. guaranteeing that a
relative value being passed in was really being completely written to,
so it didn't matter that it was relative vs specific).  Then it was
expanded to use trashing to strengthen the guarantee.

However, this created a sequence point issue for lines like:

    Val_Init_Block(out, Split_Lines(out));

Because the first argument was being processed by a function (that
"trashed" out's bits) there's no guarantee about whether the function
would run before or after the Split_Lines call.

Such calls were fairly dodgy in the first place, so it's best not to
make them.  This patches to go through a temporary variable.