Actions use SYM_XXX constants, no IS_BINARY_ACT
Historically there were two different numeric-values-for-words lists,
one was the actions e.g. A_OPEN and the other was the SYM_XXX constants.

The main barrier to merging the two was that actions were ordered in
such a way that all the "binary" math-like operators were in a common
range.  Detecting that commonality was used to do some common argument
processing in the switch statements behind the actions.

This implements that common logic in other ways just to get past the
issue, and merges the lists together.  It also gets rid of the "type
testing action numbers"--which were small integers.  Instead it uses
specialization of a common HAS-TYPE? function, the table of which
is processed in usermode code.