REBVAL .func => .paramlist, .context => .varlist
The payload structure for function and context had pointers to functions
and contexts in them, which was confusing.  This makes it clear which
piece of the function is actually being held, and uses the now-common
terminology for those things.

This also pushes a little more toward fewer macros, by having the
few initializers who require lvalue semantics just write out the
initialization fully at the callsite.