Increase REBSER size to 2xREBVAL size, cache underlying
This change is part of a shift to be able to create a new kind of
REBSER node which is actually an implicitly terminated array allocation
of length exactly 2.

The larger REBSER size is offset by getting enough extra information
in the node to be able to store WORD! hash table entries.  The two
new fields--one that has all but the low two bits available, and the
other which can hold a full pointer, will be used for other series
types for features including recording the file and line of origin.

This implements a first usage of the new "link" field in the REBSER,
in order to cache the "underlying" pointer that is needed in order to
know what size frame to push--in spite of specializations.  Without
the cache, the functions in a composition have to be pre-walked, which
was being paid for on every call.