Split REBVAL payload into "extra" vs "payload"
Having the payloads be one big union was creating a problem when there
was a common field between payloads, e.g. the binding information.
The problem is that since it was a union, there was no way to write
this field generically without risking undefined behavior, since you'd
have to pick one of the unions to write through.

This had proven to be inconvenient, causing there to be different ways
of initializing the same binding field based on which fields you wanted
to keep valid.  By splitting it into two parts, an additional advantage
of getting  natural 64-bit alignment on the payload is given--which
required padding and more explanation before.

Also in the commit is a simplification of Rebol library handles, in
order to fit their entire information into the payload--and keep from
having to have a separate "library pool".  This is done by having them
make use of the singular REBARR technique.