Remove padding from payloads for Integer, Decimal
Breaking out the "extra" was intended to remove the need for the
padding in the payloads for 64-bit alignment--but somehow that part
of the change didn't get committed, and Travis did not complain.

Also corrects the filling of the definitional "leave" native slots
from using the RETURN function, and a duplicate trace line.