Attempt to fix alignment issues on some 32-bit platforms
The change which broke REBVAL payloads into two parts was motivated by
needing to be able to set one member in a common portion of the union
without affecting the other part (the C standard says that once you
*write* one of the fields in a union, the others go invalid--though it
has some standards allowance for *reading* fields if they are common).

REBEVT and deci did not generate errors on Travis builds for some
reason, but were too large to fit in the reduced payload.  This adapts
the types so that they are themselves split across the two parts.