Share GC marking and GC managing in REBSER/REBVAL
The concept of putting two REBVALs into a REBSER node has further
motivation, so this takes another step on the path to implementing it.
Two bits in the REBVAL header are reserved.

This change was motivated by a need also to free up some info bits
for the series.  While there would be other ways of doing that, this
was on the agenda anyway.

(Also fixes zeroing of deci values for money introduced in last commit)