Speed VAL_TYPE() in Debug Build (~3x)
Performance of the debug build has not been analyzed much, but since it
has been getting slower then a run through Valgrind found that
the very-frequently-called VAL_TYPE() was unduly slow.

Although it's marked as an inline function, debug builds do not inline
(for the sake of stepping).  This is generally a good thing, but it
was calling several other functions.  While turning those functions into
macros would be bad if they repeated their arguments, the raw type
extraction did not--so that could be a macro.  This inlines
IS_TRASH_DEBUG() by hand, while leaving it as a function for other
purposes (since it repeats the argument).