Remove VALUE_FLAG_ARRAY cache
Prior to the commit that segmented the Reb_Value into a 64-bit aligned
payload and an "extra" field (containing binding), it was necessary
for value copying in specific binding to know if things were arrays or
not, quite often.  This was a frequent need in the system already, so
it seemed sensible to optimize it.

That check is no longer needed, since the binding lives in its own
field and there's no danger of disengaging the other union data by
writing a word through an array (or vice versa).  There are also now
several other types participating in specific binding.

As for the speed of the check in general beyond that--it's already
fairly optimized anyway.  The array types are in order in the list, so
it's just a range check.  They could probably be lined up to have a
common bit pattern and not even require the type to be masked or
shifted out anyway.

Given the need for value header bits as a resource, this is removed.

and all the array types could
probably be lined up in the types to have some common bit pattern to
be checked